[docs] Update README and add example scripts for EstimatorFactory usage and debugging
